refactor: delete workflow.Workflow and rename workflow.py to scenario.py
