+++ openscenario/openscenario_utility/openscenario_utility/scenario_validation.py
@@ -20,7 +20,6 @@ from typing import List
 from pkg_resources import resource_string
 from sys import exit, modules
 from openscenario_utility.conversion import convert
-from scenario_test_runner.workflow import Workflow
 
 import argparse
 import xmlschema
@@ -104,30 +103,6 @@ def validate_file(
         except:
             overall_result = False
         print(str(path) + " : " + str(overall_result))
-
-
-def validate_workflow():
-
-    schema = xmlschema.XMLSchema(
-        resource_string(__name__, "resources/OpenSCENARIO-1.2.xsd").decode("utf-8")
-    )
-
-    parser = argparse.ArgumentParser(
-        description="Validate if the Workflow Files pass the conditions checked by validators"
-    )
-    parser.add_argument("--workflow_paths", type=Path, nargs="+")
-    parser.add_argument("--validators", nargs="+")
-    args = parser.parse_args()
-
-    scenario_validators = list()
-    for validator in args.validators:
-        scenario_validators.append(getattr(modules[__name__], validator)())
-
-    for path in args.workflow_paths:
-        workflow = Workflow(path, 30.0)
-        for scenario in workflow.scenarios:
-            validate_file(scenario.path, schema, scenario_validators)
-
 
 def validate_scenario():
